my_application/requirements.dev.txt
fastapi==0.95.2
Jinja2==3.1.2
uvicorn==0.22.0
SQLAlchemy==2.0.15
aiosqlite==0.19.0
alembic==1.11.1
pydantic==1.10.8
python-dotenv==1.0.0
greenlet==2.0.2
psycopg2-binary==2.9.6
python-multipart==0.0.6
sqlalchemy-file==0.4.0
fasteners==0.18
asyncpg==0.27.0
pytest==7.3.2
httpx==0.24.1
aiofiles==23.1.0